pi-device/requirements.txt
sounddevice==0.5.1
numpy==2.2.6
websockets==16.0
aiohttp==3.11.18
qrcode==8.0
Pillow==11.1.0